requirements.txt
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.0
google-genai>=0.3.0
python-dotenv>=1.0.1
